dom/network/tests/test_tcpsocket_enabled_with_perm.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure TCPSocket permission enabled and open works with tcp-socket perm</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test to ensure TCPSocket permission being turned on enables 
  navigator.mozTCPSocket, and mozTCPSocket.open works when
  the tcp-socket permission has been granted.
**/
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [['dom.mozTCPSocket.enabled', true]]}, function() {
  SpecialPowers.pushPermissions([{type: "tcp-socket", allow: true, context: document}], runTest);
});

function runTest() {
  ok('TCPSocket' in this, "TCPSocket should be accessible if dom.mozTCPSocket.enabled is true");

  ok(new TCPSocket('localhost', 80), "TCPSocket constructor should work for content that has the tcp-socket permission");
  ok(navigator.mozTCPSocket.open('localhost', 80), "navigator.mozTCPSocket.open should work for content that has the tcp-socket permission");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
